move `store_check_result` to reporter queue
to make sure the result always exist the server side.
